requirements.txt
# PDF 처리
PyMuPDF==1.23.8
pdf2image==1.16.3

# 이미지 처리
Pillow==10.1.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# OCR 엔진
paddlepaddle==2.5.2
paddleocr==2.7.0.3
easyocr==1.7.0
pytesseract==0.3.10

# 수식 인식
pix2text==0.2.3
latex2sympy==1.9.1

# 딥러닝 프레임워크
torch==2.1.1
torchvision==0.16.1
transformers==4.36.2

# 레이아웃 분석
layoutparser==0.3.4
detectron2==0.6

# 한국어 처리
konlpy==0.6.0
jpype1==1.4.1

# 수치 계산
numpy==1.24.3
scipy==1.11.4
sympy==1.12

# 데이터 처리
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9

# 데이터베이스
sqlite3  # 기본 포함

# API 서버
fastapi==0.104.1
uvicorn==0.25.0
pydantic==2.5.2
python-multipart==0.0.6

# 유틸리티
tqdm==4.66.1
psutil==5.9.6
python-dateutil==2.8.2

# 로깅 및 설정
python-json-logger==2.0.7
python-dotenv==1.0.0

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# 개발 도구 (선택사항)
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# 추가 의존성
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.2
Jinja2==3.1.2